Moves useProducts hook to separate hooks folder
